+++ app/Http/Controllers/Back/UnitsController.php
@@ -11,39 +11,48 @@ use Yajra\DataTables\Facades\DataTables;
 class UnitsController extends Controller
 {
     public function index()
-    {                               
-        $pageNameAr = 'وحدات الأصناف';
-        $pageNameEn = 'units';
-        return view('back.units.index' , compact('pageNameAr' , 'pageNameEn'));
-    }
+    {               
+        if((userPermissions()->units_view)){
+            $pageNameAr = 'وحدات السلع والخدمات';
+            $pageNameEn = 'units';
 
-    public function create()
-    {
-        
+            return view('back.units.index' , compact('pageNameAr' , 'pageNameEn'));
+        }else{
+            return redirect('/')->with(['notAuth' => 'عذرًا، ليس لديك صلاحية لتنفيذ طلبك']);
+        }                  
     }
 
     public function store(Request $request)
     {
-        if (request()->ajax()){
-            $this->validate($request , [
-                'name' => 'required|string|unique:units,name',
-            ],[
-                'name.required' => 'إسم الوحدة مطلوب',
-                'name.string' => 'حقل إسم الوحدة يجب ان يكون من نوع نص',
-                'name.unique' => 'إسم الوحدة مستخدم من قبل',                
-            ]);
-
-            Units::create($request->all());
-        }
+        if((userPermissions()->units_create)){
+            if (request()->ajax()){
+                $this->validate($request , [
+                    'name' => 'required|string|unique:units,name',
+                ],[
+                    'name.required' => 'إسم الوحدة مطلوب',
+                    'name.string' => 'حقل إسم الوحدة يجب ان يكون من نوع نص',
+                    'name.unique' => 'إسم الوحدة مستخدم من قبل',                
+                ]);
+    
+                Units::create($request->all());
+            }
+        }else{
+            return response()->json(['notAuth' => 'عذرًا، ليس لديك صلاحية لتنفيذ طلبك']);
+        } 
     }
 
     public function edit($id)
     {
-        if(request()->ajax()){
-            $find = Units::where('id', $id)->first();
-            return response()->json($find);
-        }
-        return view('back.welcome');
+        if((userPermissions()->units_update)){
+            if(request()->ajax()){
+                $find = Units::where('id', $id)->first();
+                return response()->json($find);
+            }
+            return view('back.welcome');
+
+        }else{
+            return response()->json(['notAuth' => 'عذرًا، ليس لديك صلاحية لتنفيذ طلبك']);
+        } 
     }
 
     public function update(Request $request, $id)
@@ -65,16 +74,20 @@ class UnitsController extends Controller
 
      
     public function destroy($id){
-        $unit = DB::table('products')->where('bigUnit', $id)->orWhere('smallUnit', $id)->first();
-
-        if ($unit) {
-            return response()->json(['cannot_delete' => 'cannot_delete']);
-        }
-
-        if (!$unit) {
-            DB::table('units')->where('id', $id)->delete();
-            return response()->json(['success_delete' => 'success_delete']);
-        }
+        if((userPermissions()->units_delete)){
+            $unit = DB::table('products')->where('bigUnit', $id)->orWhere('smallUnit', $id)->first();
+    
+            if ($unit) {
+                return response()->json(['cannot_delete' => 'cannot_delete']);
+            }
+    
+            if (!$unit) {
+                DB::table('units')->where('id', $id)->delete();
+                return response()->json(['success_delete' => 'success_delete']);
+            }
+        }else{
+            return response()->json(['notAuth' => 'عذرًا، ليس لديك صلاحية لتنفيذ طلبك']);
+        } 
     }
